Add support for WOInject in JEE environments (closes #8)

+++ src/test/java/com/woinject/AbstractInjectableTestCase.java
@@ -61,7 +61,9 @@ public abstract class AbstractInjectableTestCase {
 
     @After
     public void tearDown() {
-	application.terminate();
+	if (application != null) {
+	    application.terminate();
+	}
 
 	application = null;
     }
